Move PatchOperation type to src/types/index.ts
Per repository conventions, type definitions live in src/types/index.ts
while src/api/endpoints.ts only contains endpoint functions. Moved the
PatchOperation declaration to the shared types module and updated
consumers to import it from @/types. endpoints.ts re-exports the type
for backward compatibility.
